Adds check (safety but not very nice - yet) to prevent errors when GIT_REV constant doesn't exist
Anyway. It's to be discussed/re-thought if this is really a constant or more a config value?! At the moment I would tend more to it being a config entry but not sure as it feels wrong too.

But as the enviro version is also set there it should be ok!