Add `pyproject.toml` and `poetry.lock` for dependency management
With this setup most code completion is available and matches (more or less) the currently used / required MicroPaython version (1.25.x).

Nevertheless, the Pimoroni stubs for the Pico W Enviro are still on a lower version (1.22.x) - check regularly for updates as it might be a little different as the used MicroPython version!